+++ custom_types/map_12_tasks.h
@@ -1,35 +1,36 @@
 #include <map>
 #include "my_types.h"
+#include <cinttypes>
 
 using namespace std;
 
 
-typedef map<unsigned short, bool> mt1;
-typedef map<unsigned short, mt1> mt2;
-typedef map<unsigned short, mt2> mt3;
-typedef map<unsigned short, mt3> mt4;
-typedef map<unsigned short, mt4> mt5;
-typedef map<unsigned short, mt5> mt6;
-typedef map<unsigned short, mt6> mt7;
-typedef map<unsigned short, mt7> mt8;
-typedef map<unsigned short, mt8> mt9;
-typedef map<unsigned short, mt9> mt10;
-typedef map<unsigned short, mt10> mt11;
-typedef map<unsigned short, mt11> mt12;
-typedef map<unsigned short, mt12> mt13;
-typedef map<unsigned short, mt13> mt14;
-typedef map<unsigned short, mt14> mt15;
-typedef map<unsigned short, mt15> mt16;
-typedef map<unsigned short, mt16> mt17;
-typedef map<unsigned short, mt17> mt18;
-typedef map<unsigned short, mt18> mt19;
-typedef map<unsigned short, mt19> mt20;
-typedef map<unsigned short, mt20> mt21;
-typedef map<unsigned short, mt21> mt22;
-typedef map<unsigned short, mt22> mt23;
-typedef map<unsigned short, mt23> mt24;
-typedef map<unsigned short, mt24> mt25;
-typedef map<unsigned short, mt25> mt26;
+typedef map<uint8_t, bool> mt1;
+typedef map<uint8_t, mt1> mt2;
+typedef map<uint8_t, mt2> mt3;
+typedef map<uint8_t, mt3> mt4;
+typedef map<uint8_t, mt4> mt5;
+typedef map<uint8_t, mt5> mt6;
+typedef map<uint8_t, mt6> mt7;
+typedef map<uint8_t, mt7> mt8;
+typedef map<uint8_t, mt8> mt9;
+typedef map<uint8_t, mt9> mt10;
+typedef map<uint8_t, mt10> mt11;
+typedef map<uint8_t, mt11> mt12;
+typedef map<uint8_t, mt12> mt13;
+typedef map<uint8_t, mt13> mt14;
+typedef map<uint8_t, mt14> mt15;
+typedef map<uint8_t, mt15> mt16;
+typedef map<uint8_t, mt16> mt17;
+typedef map<uint8_t, mt17> mt18;
+typedef map<uint8_t, mt18> mt19;
+typedef map<uint8_t, mt19> mt20;
+typedef map<uint8_t, mt20> mt21;
+typedef map<uint8_t, mt21> mt22;
+typedef map<uint8_t, mt22> mt23;
+typedef map<uint8_t, mt23> mt24;
+typedef map<uint8_t, mt24> mt25;
+typedef map<uint8_t, mt25> mt26;
 typedef map<my_bitset, mt26, cmpOfBitsets> mt27;
 typedef map<my_bitset, mt27, cmpOfBitsets> mt28;
 typedef map<my_bitset, mt28, cmpOfBitsets> mt29;
