+++ custom_types/map_6_tasks.h
@@ -1,22 +1,23 @@
 #include <map>
 #include "my_types.h"
+#include <cinttypes>
 
 using namespace std;
 
-typedef map<unsigned short, bool> mt1;
-typedef map<unsigned short, mt1> mt2;
-typedef map<unsigned short, mt2> mt3;
-typedef map<unsigned short, mt3> mt4;
-typedef map<unsigned short, mt4> mt5;
-typedef map<unsigned short, mt5> mt6;
-typedef map<unsigned short, mt6> mt7;
-typedef map<unsigned short, mt7> mt8;
-typedef map<unsigned short, mt8> mt9;
-typedef map<unsigned short, mt9> mt10;
-typedef map<unsigned short, mt10> mt11;
-typedef map<unsigned short, mt11> mt12;
-typedef map<unsigned short, mt12> mt13;
-typedef map<unsigned short, mt13> mt14;
+typedef map<uint8_t, bool> mt1;
+typedef map<uint8_t, mt1> mt2;
+typedef map<uint8_t, mt2> mt3;
+typedef map<uint8_t, mt3> mt4;
+typedef map<uint8_t, mt4> mt5;
+typedef map<uint8_t, mt5> mt6;
+typedef map<uint8_t, mt6> mt7;
+typedef map<uint8_t, mt7> mt8;
+typedef map<uint8_t, mt8> mt9;
+typedef map<uint8_t, mt9> mt10;
+typedef map<uint8_t, mt10> mt11;
+typedef map<uint8_t, mt11> mt12;
+typedef map<uint8_t, mt12> mt13;
+typedef map<uint8_t, mt13> mt14;
 typedef map<my_bitset, mt14, cmpOfBitsets> mt15;
 typedef map<my_bitset, mt15, cmpOfBitsets> mt16;
 typedef map<my_bitset, mt16, cmpOfBitsets> mt17;
