+++ custom_types/map_8_tasks.h
@@ -1,27 +1,28 @@
 #include <map>
 #include "my_types.h"
+#include <cinttypes>
 
 using namespace std;
 
 
-typedef map<unsigned short, bool> mt1;
-typedef map<unsigned short, mt1> mt2;
-typedef map<unsigned short, mt2> mt3;
-typedef map<unsigned short, mt3> mt4;
-typedef map<unsigned short, mt4> mt5;
-typedef map<unsigned short, mt5> mt6;
-typedef map<unsigned short, mt6> mt7;
-typedef map<unsigned short, mt7> mt8;
-typedef map<unsigned short, mt8> mt9;
-typedef map<unsigned short, mt9> mt10;
-typedef map<unsigned short, mt10> mt11;
-typedef map<unsigned short, mt11> mt12;
-typedef map<unsigned short, mt12> mt13;
-typedef map<unsigned short, mt13> mt14;
-typedef map<unsigned short, mt14> mt15;
-typedef map<unsigned short, mt15> mt16;
-typedef map<unsigned short, mt16> mt17;
-typedef map<unsigned short, mt17> mt18;
+typedef map<uint8_t, bool> mt1;
+typedef map<uint8_t, mt1> mt2;
+typedef map<uint8_t, mt2> mt3;
+typedef map<uint8_t, mt3> mt4;
+typedef map<uint8_t, mt4> mt5;
+typedef map<uint8_t, mt5> mt6;
+typedef map<uint8_t, mt6> mt7;
+typedef map<uint8_t, mt7> mt8;
+typedef map<uint8_t, mt8> mt9;
+typedef map<uint8_t, mt9> mt10;
+typedef map<uint8_t, mt10> mt11;
+typedef map<uint8_t, mt11> mt12;
+typedef map<uint8_t, mt12> mt13;
+typedef map<uint8_t, mt13> mt14;
+typedef map<uint8_t, mt14> mt15;
+typedef map<uint8_t, mt15> mt16;
+typedef map<uint8_t, mt16> mt17;
+typedef map<uint8_t, mt17> mt18;
 typedef map<my_bitset, mt18, cmpOfBitsets> mt19;
 typedef map<my_bitset, mt19, cmpOfBitsets> mt20;
 typedef map<my_bitset, mt20, cmpOfBitsets> mt21;
