+++ get_delta_t.h
@@ -1,3 +1,3 @@
-unsigned short get_dt_to_next_release(const state&, const unsigned short int, const unsigned short);
-unsigned short get_dt_to_next_completion(const state&, const unsigned short int, const unsigned short*);
-unsigned short deltaT(const state&, const unsigned short int, const unsigned short int, const unsigned short*);
+uint8_t get_dt_to_next_release(const state&, const uint8_t, const uint8_t);
+uint8_t get_dt_to_next_completion(const state&, const uint8_t, const uint8_t*);
+uint8_t deltaT(const state&, const uint8_t, const uint8_t, const uint8_t*);
