+++ scheduler.h
@@ -1,1 +1,1 @@
-void sortTasksByPriorities(const state&, const unsigned short int, unsigned short*);
+void sortTasksByPriorities(const state&, const uint8_t, uint8_t*);
